backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
google-genai
python-dotenv
